backend/requirements.txt
requests
fastapi
uvicorn
elasticsearch~=9.0.2
python-dotenv
sqlalchemy
pymysql
pymongo
boto3
cryptography
tqdm
openai
scipy
python-multipart
passlib[bcrypt]
email-validator
pillow